core: Move unstable::exchange_alloc to rt::global_heap
