+++ src/libcore/rt/global_heap.rs
